Move lazy type alias checks to non-hir-wfck
