+++ tests/ui/lazy-type-alias/inherent-impls-overflow.rs
@@ -16,12 +16,12 @@ impl Loop {}
 
 type Poly0<T> = Poly1<(T,)>;
 //[current]~^ ERROR overflow normalizing the type alias `Poly0<(((((((...,),),),),),),)>`
-//~^^ ERROR type parameter `T` is only used recursively
-//[next]~^^^ ERROR overflow evaluating the requirement
+//[next]~^^ ERROR type parameter `T` is only used recursively
+//[next]~| ERROR overflow evaluating the requirement
 type Poly1<T> = Poly0<(T,)>;
 //[current]~^ ERROR  overflow normalizing the type alias `Poly1<(((((((...,),),),),),),)>`
-//~^^ ERROR type parameter `T` is only used recursively
-//[next]~^^^ ERROR overflow evaluating the requirement
+//[next]~^^ ERROR type parameter `T` is only used recursively
+//[next]~| ERROR overflow evaluating the requirement
 
 impl Poly0<()> {}
 //[current]~^ ERROR overflow normalizing the type alias `Poly1<(((((((...,),),),),),),)>`
