+++ tests/ui/lazy-type-alias/unconstrained-params-in-impl-due-to-overflow.stderr
@@ -4,26 +4,15 @@ error[E0207]: the type parameter `T` is not constrained by the impl trait, self 
 LL | impl<T> Loop<T> {}
    |      ^ unconstrained type parameter
 
-error: type parameter `T` is only used recursively
-  --> $DIR/unconstrained-params-in-impl-due-to-overflow.rs:6:21
-   |
-LL | type Loop<T> = Loop<T>;
-   |           -         ^
-   |           |
-   |           type parameter must be used non-recursively in the definition
-   |
-   = help: consider removing `T` or referring to it in the body of the type alias
-   = note: all type parameters must be used in a non-recursive way in order to constrain their variance
-
 error[E0275]: overflow normalizing the type alias `Loop<T>`
-  --> $DIR/unconstrained-params-in-impl-due-to-overflow.rs:6:16
+  --> $DIR/unconstrained-params-in-impl-due-to-overflow.rs:6:1
    |
 LL | type Loop<T> = Loop<T>;
-   |                ^^^^^^^
+   | ^^^^^^^^^^^^
    |
    = note: in case this is a recursive type alias, consider using a struct, enum, or union instead
 
-error: aborting due to 3 previous errors
+error: aborting due to 2 previous errors
 
 Some errors have detailed explanations: E0207, E0275.
 For more information about an error, try `rustc --explain E0207`.
